125 - Added internal DayCycle state to spawner, added despawn method.
ZombieSpawner now acts as a listener to DayController, and will react to the different DayCycle state it is pushed.
Added a Despawn Method that destroys the GameObjects of type Zombie. Will refactor to IEnumerator to be able to yield for some amount of time so that all zombies aren't instantly destroyed.